+++ app/controllers/guests_controller.rb
@@ -7,13 +7,13 @@ class GuestsController < ApplicationController
   def index
     @guest_form = GuestForm.new(guest_form_params)
 
-    departure_info = RequestGeocodeService.new(@guest_form).call
+    departure_info = Api::GeocodeService.new(@guest_form).call
     if departure_info.include?(:error)
       flash.now[:error] = departure_info[:error]
       return render :new, status: :unprocessable_entity
     end
 
-    @results = RequestNearbyService.new(departure_info, @guest_form).call
+    @results = Api::NearbyService.new(departure_info, @guest_form).call
     if @results.include?(:error)
       flash.now[:error] = @results[:error]
       return render :new, status: :unprocessable_entity
