+++ app/controllers/searches_controller.rb
@@ -27,7 +27,7 @@ class SearchesController < ApplicationController
 
   def create
     @search_term_form = SearchTermForm.new(search_term_form_params)
-    results = RequestNearbyService.new(@departure_info, @search_term_form).call
+    results = Api::NearbyService.new(@departure_info, @search_term_form).call
 
     if results.include?(:error)
       flash.now[:error] = results[:error]
