+++ app/services/api/request.rb
@@ -1,0 +1,11 @@
+module Api
+  module Request
+    private
+
+    def send_request(url)
+      uri = URI.parse(url)
+      json = Net::HTTP.get(uri)
+      JSON.parse(json, symbolize_names: true)
+    end
+  end
+end
