--- app/services/request_api.rb
@@ -1,9 +1,0 @@
-module RequestApi
-  private
-
-  def send_request(url)
-    uri = URI.parse(url)
-    json = Net::HTTP.get(uri)
-    JSON.parse(json, symbolize_names: true)
-  end
-end
